separated components from client info into different files
